+++ lib/src/commands.dart
@@ -35,7 +35,10 @@ enum GYWControlCode {
   setFont(0x08),
 
   /// Enable or disable screen auto-rotate
-  autoRotateScreen(0x0A);
+  autoRotateScreen(0x0A),
+
+  /// Enable or disable the display backlight
+  enableBacklight(0x0B);
 
   /// Control code value used internally on the device
   final int value;