+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * [ADD] Add size and color parameters to TextDrawing
 * [ADD] Convert WhiteScreen into BlankScreen with a color parameter
 * [FIX] Upgrade flutter_blue_plus version
+* [ADD] Add a method to enable or disable the display backlight
 * [ADD] Add a method to read the firmware version
 
 ## 1.1.1